Move StandardsPage to folder
